requirements-deploy.txt
beautifulsoup4>=4.13.0
flask>=3.0.0
flask-cors>=4.0.0
flask-jwt-extended>=4.6.0
flask-limiter>=3.5.0
flask-login>=0.6.2
flask-sqlalchemy>=3.1.0
flask-wtf>=1.1.1
google-api-python-client>=2.100.0
google-auth>=2.22.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
matplotlib>=3.7.2
numpy>=1.24.3
oauthlib>=3.2.2
openai>=1.12.0
pillow>=10.0.0
psycopg2-binary>=2.9.7
requests>=2.31.0
schedule>=1.2.0
sendgrid>=6.10.0
stripe>=7.5.0
trafilatura>=1.5.0
twilio>=8.5.0